interpolation of facsimilies
In the case of multiple primitive neume componenets recognized within a single CC, this will attempt to guess where the bounding boxes for each separate neume component are. This requires the neumes passed in to be of the form neume.name.contour.name.contour.etc... For the meantime, this will offer some way of recognizing how well the pitch classification is working on non primitive CC. Also fixed an issue where no non-neume glyphs were being generated.